Converted floating point membe varables to double percision.

+++ release1.1/src/plugins/coremodules/INPUT_REC.h
@@ -4,13 +4,13 @@
 class  INPUT_REC { // climate-data input records
 public:
 
-    float S_n;	// net solar radiation (W/m^2)
-    float I_lw;	// incoming longwave (thermal) rad (W/m^2)
-    float T_a;	// air temp (C)
-    float e_a;	// vapor pressure (Pa)
-    float u;	// wind speed (m/sec)
-    float T_g;	// soil temp at depth z_g (C)
-    float F_g;	// soil flux at depth z_g (W/m^2)
+    double S_n;	// net solar radiation (W/m^2)
+    double I_lw;	// incoming longwave (thermal) rad (W/m^2)
+    double T_a;	// air temp (C)
+    double e_a;	// vapor pressure (Pa)
+    double u;	// wind speed (m/sec)
+    double T_g;	// soil temp at depth z_g (C)
+    double F_g;	// soil flux at depth z_g (W/m^2)
 };
 
 #endif
