+++ release1.1/src/plugins/coremodules/PRECIP_REC.h
@@ -4,12 +4,12 @@
 class  PRECIP_REC { // precip-data input records
 public:
 
-    float	  m_pp;		// total precipitation mass (kg/m^2)
-    float	  m_rain;	// mass of rain in precip   (kg/m^2)
-    float	  m_snow;	// mass of snow in precip   (kg/m^2)
-    float	  m_drift;	// mass of snow drift       (kg/m^2)
-    float	  m_subl;	// mass of snow drift       (kg/m^2)
-    float	  z_snow;	// depth of snow in precip  (m)
+    double	  m_pp;		// total precipitation mass (kg/m^2)
+    double	  m_rain;	// mass of rain in precip   (kg/m^2)
+    double	  m_snow;	// mass of snow in precip   (kg/m^2)
+    double	  m_drift;	// mass of snow drift       (kg/m^2)
+    double	  m_subl;	// mass of snow drift       (kg/m^2)
+    double	  z_snow;	// depth of snow in precip  (m)
 };
 
 #endif
